Fix a bunch of horrible memory errors
C is a programming language

+++ src/frontend/sndprint.h
@@ -4,7 +4,7 @@
 void snd_configure(char*, int, int, double, double, int);
 void snd_init(int);
 void snd_close(void);
-int snd_send(double, int, double);
+int snd_send(double, int, float);
 int snd_format(char*);
 double snd_get_samplerate(void);
